add ability to keep boostrap.py

+++ cctbx/update_bootstrap.sh
@@ -3,12 +3,18 @@
 # this () { echo $(readlink -f $(dirname ${BASH_SOURCE[0]})); }
 shopt -s expand_aliases
 alias this="readlink -f \$(dirname \${BASH_SOURCE[0]})"
-
+export BOOSTRAP_SOURCE="https://raw.githubusercontent.com/cctbx/cctbx_project/master/libtbx/auto_build/bootstrap.py"
 
 pushd $(this)
 if [[ -f bootstrap.py ]]
 then
-    rm bootstrap.py
+    if [[ -z ${LEAVE_BOOSTRAP+x} ]]
+    then
+        rm bootstrap.py
+    fi
 fi
-wget "https://raw.githubusercontent.com/cctbx/cctbx_project/master/libtbx/auto_build/bootstrap.py" --no-check-certificate
+if [[ ! -e bootstrap.py ]]
+then
+    wget $BOOSTRAP_SOURCE --no-check-certificate
+fi
 popd
